backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.20.0
cobble==0.1.4
confection==0.1.5
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.0
cymem==2.0.10
daphne==4.1.2
datasets==3.1.0
decorator==5.1.1
diff-match-patch==20241021
dill==0.3.8
Django==4.2.17
django-allauth==65.0.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-import-export==4.2.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docx==0.2.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
ffmpeg-python==0.2.0
filelock==3.16.1
flashtext==2.7
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2024.9.0
future==0.18.3
google-api-core==2.24.0
google-api-python-client==2.113.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.66.0
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.26.5
hyperlink==21.0.0
idna==3.8
imageio==2.36.1
imageio-ffmpeg==0.5.1
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.2.0
kombu==5.4.1
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
mammoth==1.8.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mediawikiapi==1.2
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
networkx==3.1
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
pke @ git+https://github.com/boudinfl/pke.git@69871ffdb720b83df23684fea53ec8776fd87e63
preshed==3.0.9
proglog==0.1.10
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psycopg2-binary==2.9.9
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.25.1
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyperclip==1.9.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
PyYAML==6.0.2
redis==5.0.8
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sense2vec==2.0.2
sentencepiece==0.2.0
service-identity==24.1.0
setuptools==73.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.12.0
sqlparse==0.5.1
srsly==2.5.0
strsim==0.0.3
sympy==1.13.1
tablib==3.7.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.1
Twisted==24.7.0
txaio==23.1.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zope.interface==7.0.1
